sparse: Silence logging from droid-reclaim-memory

+++ sparse/usr/bin/droid/droid-reclaim-memory.sh
@@ -4,10 +4,12 @@ while [ "$(wc -l < /proc/swaps)" -lt 2 ]; do
     sleep 1
 done
 
-TASKS=$(cat /sys/fs/cgroup/systemd/system.slice/*/cgroup.procs /sys/fs/cgroup/unified/system.slice/*/cgroup.procs /sys/fs/cgroup/system.slice/*/cgroup.procs)
+TASKS=$(find /sys/fs/cgroup -regex '.*system.slice.*cgroup.procs' -exec cat {} + | sort -u)
 
 for task in $TASKS; do
-    echo "all" > /proc/"$task"/reclaim
+    if [ -f /proc/"$task"/reclaim ]; then
+        echo "all" > /proc/"$task"/reclaim
+    fi
 done
 
 DONE_FILE="/tmp/.droid-reclaim-memory-ran"
@@ -15,10 +17,12 @@ DONE_FILE="/tmp/.droid-reclaim-memory-ran"
 if [ ! -f $DONE_FILE ]; then
     # right after first login, no user apps are running yet
     # reclaim also from the user session.
-    TASKS=$(cat /sys/fs/cgroup/systemd/user.slice/user-*.slice/*/cgroup.procs /sys/fs/cgroup/unified/user.slice/user-*.slice/*/cgroup.procs /sys/fs/cgroup/user.slice/user-*.slice/*/cgroup.procs)
+    TASKS=$(find /sys/fs/cgroup -regex '.*user.slice.*cgroup.procs' -exec cat {} + | sort -u)
 
     for task in $TASKS; do
-        echo "all" > /proc/"$task"/reclaim
+        if [ -f /proc/"$task"/reclaim ]; then
+            echo "all" > /proc/"$task"/reclaim
+        fi
     done
 
     touch $DONE_FILE
